Rename Base class and files into FederatedClient and DecentralizedClient
