+++ discojs/src/client/decentralized/decentralized_client.ts
@@ -18,7 +18,7 @@ const debug = createDebug("discojs:client:decentralized");
  * with the server is based off regular WebSockets, whereas peer-to-peer communication uses
  * WebRTC for Node.js.
  */
-export class Base extends Client {
+export class DecentralizedClient extends Client {
   /**
    * The pool of peers to communicate with during the current training round.
    */
