+++ discojs/src/client/decentralized/index.ts
@@ -1,2 +1,2 @@
-export { Base as DecentralizedClient } from './base.js'
+export { DecentralizedClient } from './decentralized_client.js'
 export * as messages from './messages.js'
